+++ bin/generate_plugins.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import argparse
+import sys
 from importlib import import_module
 from itertools import chain
 from pathlib import Path
@@ -7,6 +8,9 @@ from textwrap import dedent
 
 import inspectortiger.plugins
 from inspectortiger import Inspector
+
+sys.path.insert(0, "../inspectortiger")  # i dont like this either
+
 
 BASE = Path(inspectortiger.plugins.__file__).parent